Move evaluation interface to `Eval.hs`.
* `analyzeCheck`
* `analyzeInvariants`
* `runInvariantAnalysis`
* `runPropertyAnalysis`
